[feature/SBOT-79]: [feat] 특정 게시물 조회 기초 틀 구현

+++ common/src/commonMain/kotlin/common/model/reseponse/post/Post.kt
@@ -13,7 +13,7 @@ data class Post(
     @SerialName("author") val author: User,
     @SerialName("votes") val votes: List<Vote>,
     @SerialName("categories") val categories: List<Category>,
-    @SerialName("selected_vote") val selectedVote: Long?,
+    @SerialName("selected_vote") val selectedVote: Int?,
     @SerialName("is_liked") val isLiked: Boolean?,
     @SerialName("created_at") val createdAt: String,
     @SerialName("updated_at") val updatedAt: String?
